feat: init lighthouse
