+++ lighthouse/const/index.js
@@ -1,0 +1,6 @@
+const apiKey = "acd55174-7003-4cd0-93b8-7cc7d3a5b3c1";
+const publicKey = "0xf6f1058c985572422b6ed0B5Fd4Af2Ba2704922B";
+const privateKey =
+    "f80f8f01a2a34a6977d09f84ecf03eadcd4a4f73a52cee71218504964c6c33dc";
+
+module.exports = {apiKey, publicKey, privateKey};
