+++ lighthouse/utils/signAuthMessage.js
